% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{A}
\alias{A}
\title{Compute CO2 assimilation for a leaf given parameters and environment}
\usage{
A(
  pars = parameters(),
  Np = 0.1,
  f_Nc = 0.15,
  f_Nr = 0.3,
  VPD = 1,
  PAR = 1000,
  Tleaf = 25,
  Ca = 400
)
}
\arguments{
\item{pars}{Parameters as generated by \code{\link{parameters}}}

\item{Np}{Photosynthetic nitrogen content (mol/m2)}

\item{f_Nc}{Fraction of photosynthetic nitrogen allocated to chlorophyll}

\item{f_Nr}{Fraction of photosynthetic nitrogen allocated to Rubisco}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{PAR}{Photosynthetically active radiation (umol/m2/s)}

\item{Tleaf}{Leaf temperature (Celsius)}

\item{Ca}{CO2 molar fraction in the air (umol/mol)}
}
\value{
A named vector with gross CO2 assimilation (\code{"Ag"}), net CO2 assimilation (\code{"A"}),
net CO2 assimilation limited by carboxylation capacity (\code{"Ac"}) and
net CO2 assimilation limited by electron transport capacity (\code{"Aj"}),
excess energy as potential assimilation (\code{"E}) and the relative excess energy
(\code{"rE}).
}
\description{
Compute CO2 assimilation for a leaf given parameters and environment
}
\examples{
pars = parameters()
As = A(pars)
As["A"] == min(As["Ac"], As["Aj"])
}
