% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{Acan}
\alias{Acan}
\title{Calculate canopy photosynthesis and intermediate variables per layer}
\usage{
Acan(can, env, pars, vars)
}
\arguments{
\item{can}{Canopy structure object as generated by \code{\link{canopy}}}

\item{env}{Object containing environmental conditions at a particular time as
generated by \code{\link{interpolate_meteo}}}

\item{pars}{Parameters as generated by \code{\link{parameters}}}

\item{vars}{Variables as generated by \code{\link{variables}}}
}
\value{
List with the following information:
\describe{
\item{Acan}{Total canopy gross CO2 assimilation per unit of ground (mumol/m2/s)}
\item{Alayer}{Canopy gross CO2 assimilation of each layer per unit of ground (mumol/m2/s)}
\item{Ilayer}{Average PAR intercepted by each layer per unit of ground (mumol/m2/s)}
\item{fsun}{Fraction of leaves that are sunlit in each layer}
\item{Ashade}{Average net CO2 assimilation per unit of shaded leaf area in each layer (mumol/m2/s)}
\item{Asun}{Average net CO2 assimilation per unit of sunlit leaf area in each layer (mumol/m2/s)}
}
}
\description{
Calculate canopy photosynthesis and intermediate variables per layer
}
\examples{
pars = parameters()
vars = variables(pars)
can = canopy(pars, vars, 15)
weather = DailyWeather()
env = interpolate_meteo(weather, weather.DL/2)
can_fluxes = Acan(can, env, pars, vars)
plot(can$Lm, can_fluxes$Alayer)
}
