% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{canopy}
\alias{canopy}
\title{Canopy structure in layers}
\usage{
canopy(pars, vars, nlayers)
}
\arguments{
\item{pars}{Parameters as generated by \code{\link{parameters}}}

\item{vars}{Variables as generated by \code{\link{variables}}}

\item{nlayers}{Number of layers into which the canopy should be divided}
}
\value{
A Julia object with photosyntetic traits per layer as well as pre-calculated
values required for computing PAR distribution
\describe{
\item{DeltaL}{Increment of leaf area index of each layer}
\item{Lm}{Cumulative LAI from the top of the canopy to the middle of each layer}
\item{Lu}{Cumulative LAI from the top of the canopy to the upper boundary of each layer}
\item{Fu}{Matrix to compute fraction of upward scattered PAR that is intercepted}
\item{Fd}{Matrix to compute fraction of downward scattered PAR that is intercepted}
\item{fdif}{Fraction of diffuse PAR intercepted by each layer}
\item{Vcmax25}{Temperature-normalized maximum rate of carboxylation in each layer}
\item{Rd25}{Temperature-normalized of mitochondrial respiration in each layer}
\item{Jmax25}{Temperature-normalized maximum rate of electron transport in each layer}
\item{alpha}{Leaf absorptance in each layer}
\item{sigma}{Leaf scattering coefficient in each layer}
}
}
\description{
Canopy structure in layers
}
\examples{
pars = parameters()
vars = variables(pars)
can = canopy(pars, vars, 15)
}
