% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{DailyWeather}
\alias{DailyWeather}
\title{Combine weather information at the daily level}
\usage{
DailyWeather(
  lat = 52,
  doy = 182L,
  tau = 0.75,
  Tmin = 20,
  Tmax = 30,
  RH = 0.4,
  Ca = 400
)
}
\arguments{
\item{lat}{Latitude of the location in hexadecimal degrees}

\item{doy}{Day of the year being simulated}

\item{tau}{Atmospheric transmissivity}

\item{Tmin}{Minimum temperature of the day (Celsius)}

\item{Tmax}{Maximum temperature of the day (Celsius)}

\item{RH}{Relative humidity at sunrise}

\item{Ca}{CO2 molar fraction in the air (umol/mol)}
}
\value{
A Julia object collecting the weather variables after some transformations:
\describe{
\item{lat}{Latitude of the location (rad)}
\item{doy}{Day of the year being simulated}
\item{tsr}{Time of the day at which sunrise occurs (h)}
\item{DL}{Diurnal length, that is, difference between sunrise and sunset (h)}
\item{tau}{Atmospheric transmissivity}
\item{Temp}{Function to interpolate temperature during the day (does not work in R)}
\item{ea}{Air vapour pressure at sunrise (kPa)}
\item{Ca}{CO2 molar fraction in the air (umol/mol)}
}
}
\description{
Combine weather information at the daily level
}
\examples{
weather = DailyWeather()
weather$DL
}
