% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{Ecan}
\alias{Ecan}
\title{Calculate excess energy per layer}
\usage{
Ecan(can, env, pars, vars)
}
\arguments{
\item{can}{Canopy structure object as generated by \code{\link{canopy}}}

\item{env}{Object containing environmental conditions at a particular time as
generated by \code{\link{interpolate_meteo}}}

\item{pars}{Parameters as generated by \code{\link{parameters}}}

\item{vars}{Variables as generated by \code{\link{variables}}}
}
\value{
List with the following information:
\describe{
\item{Ecan}{Total canopy excess energy as potential assimilation (mumol/m2/s)}
\item{Elayer}{Excess energy as of each layer per unit of ground (mumol/m2/s)}
\item{Eshade}{Excess energy as in shade per unit of shaded leaf area in each layer (mumol/m2/s)}
\item{Esun}{Excess energy as in sunlit per unit of shaded leaf area in each layer (mumol/m2/s)}
}
}
\description{
Calculate excess energy per layer
}
\examples{
pars = parameters()
vars = variables(pars)
can = canopy(pars, vars, 15)
weather = DailyWeather()
env = interpolate_meteo(weather, weather.DL/2)
can_fluxes = Acan(can, env, pars, vars)
plot(can$Lm, can_fluxes$Elayer)
}
