% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{Ediurnal}
\alias{Ediurnal}
\title{Calculate diurnal excess of energy}
\usage{
Ediurnal(weather, can, pars, vars, nsteps = 3, ngrid = 5)
}
\arguments{
\item{weather}{Object with daily weather variables as generated by \code{\link{DailyWeather}}}

\item{can}{Canopy structure object as generated by \code{\link{canopy}}}

\item{pars}{Parameters as generated by \code{\link{parameters}}}

\item{vars}{Variables as generated by \code{\link{variables}}}

\item{nsteps}{Numbers of segments in which the day should be subdivided for numerical integration}

\item{ngrid}{Order of the Gaussian-Legendre integration to be performed within each segment}
}
\value{

}
\description{
Calculate diurnal excess of energy
}
\examples{
weather = DailyWeather()
pars = parameters()
vars = variables(pars)
can = canopy(pars, vars, 15)
Eday = Ediurnal(weather, can, pars, vars)
}
