% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{ellipsoidal}
\alias{ellipsoidal}
\title{Ellipsoidal leaf angle distribution}
\usage{
ellipsoidal(X, angles)
}
\arguments{
\item{X}{Ellipsoidal parameters (must have positive value)}

\item{angles}{Leaf inclination angles (radians)}
}
\value{
Probability density of the leaf inclination angle(s)
}
\description{
Ellipsoidal leaf angle distribution
}
\examples{
curve(ellipsoidal(2.5, x), 0, pi/2)
}
