% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{Ican}
\alias{Ican}
\title{Calculate canopy photosynthesis per layer}
\usage{
Ican(can, env, pars, vars)
}
\arguments{
\item{can}{Canopy structure object as generated by \code{\link{canopy}}}

\item{env}{Object containing environmental conditions at a particular time as
generated by \code{\link{interpolate_meteo}}}

\item{pars}{Parameters as generated by \code{\link{parameters}}}

\item{vars}{Variables as generated by \code{\link{variables}}}
}
\value{
List with the following information:
\describe{
\item{Ican}{Total PAR intercepted by canopy per unit of ground (mumol/m2/s)}
\item{Ilayer}{PAR intercepted by each layer per unit of ground (mumol/m2/s)}
}
}
\description{
Calculate canopy photosynthesis per layer
}
\examples{
pars = parameters()
vars = variables(pars)
can = canopy(pars, vars, 15)
weather = DailyWeather()
env = interpolate_meteo(weather, weather.DL/2)
can_fluxes = Ican(can, env, pars, vars)
plot(can$Lm, can_fluxes$Ilayer)
}
