% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{parameters}
\alias{parameters}
\title{Specify parameters of the virtual crop}
\usage{
parameters(
  Sco25 = 2800,
  E_Sco = -24.46,
  Kmc25 = 270,
  E_Kmc = 80.99,
  Kmo25 = 165,
  E_Kmo = 23.72,
  ar = 3.5/786,
  E_Vcmax = 65.33,
  theta = 0.7,
  Knc = 0.076/25,
  Phi2LL = 0.85,
  Phi1LL = 1,
  fcyc = 0.1,
  Topt_Phi2 = 22.5 + 273.15,
  Omega = 36.5,
  aj = 0.0159,
  ks = 125,
  E_Jmax = 30,
  D_Jmax = 200,
  S_Jmax = 0.65,
  E_Rd = 46.39,
  f_Rd = 0.01,
  gm25 = 0.4,
  E_gm = 70.2,
  S_gm = 0.32,
  D_gm = 94,
  gs0 = 0.05/1.56,
  sgs = 3,
  D0 = 1.5,
  gb = 0.5,
  angles = list(X = 1, rtol = 0.001),
  Ncmin = 0.0084,
  sigma_soil = 0.21,
  sky = list(mode = "standard", nsky = 10),
  lat = 52,
  Ca = 400
)
}
\arguments{
\item{Sco25}{Rubisco specificity factor at 25 C}

\item{E_Sco}{Apparent activation energy of Sc/o (kJ/mol)}

\item{Kmc25}{Michaelis-Menten constant of carboxylation with respect to CO2 at 25 C (mol/mol)}

\item{E_Kmc}{Activation energy of Kmc (kJ/mol)}

\item{Kmo25}{Michaelis-Menten constant of oxygenation with respect to O2 at 25 C (mol/mol)}

\item{E_Kmo}{Activation energy of Kmo (kJ/mol)}

\item{ar}{Ratio between Vcmax25 and Nr (1/s)}

\item{E_Vcmax}{Activation energy of Vcmax (kJ/mol)}

\item{theta}{Curvature parameter in the light response of the electron transport}

\item{Knc}{Nc at which the leaf absorbs 50\% of incoming PAR (mol/m2)}

\item{Phi2LL}{PSII quantum yield at low light}

\item{Phi1LL}{PSI quantum yield at low light}

\item{fcyc}{Fraction of cyclic electron transport}

\item{Topt_Phi2}{Optimal temperature for PSII quantum yield}

\item{Omega}{sigma/√2 in the Gaussian temperature function for PSII quantum yield}

\item{aj}{Ratio between Jmax25 and Nt (1/s)}

\item{ks}{Ratio between Ns and Jmax25 (s)}

\item{E_Jmax}{Activation energy Jmax (kJ/mol)}

\item{D_Jmax}{Deactivation energy of Jmax (kJ/mol)}

\item{S_Jmax}{Entropy coefficient of Jmax (kJ/mol/K)}

\item{E_Rd}{Activation energy of Rd (kJ/mol)}

\item{f_Rd}{Ratio between Rd25 and Vcmax25}

\item{gm25}{Mesophyll conductance (mol/m2/s)}

\item{E_gm}{Activation energy gm (kJ/mol)}

\item{S_gm}{Entropy coefficient of gm (kJ/mol/K)}

\item{D_gm}{Deactivation energy of gm (kJ/mol)}

\item{gs0}{Minimum stomatal conductance (mol/m2/s)}

\item{sgs}{Scaling factor between gross assimilation and stomatal conductance}

\item{D0}{Sensitivity of gs to VPD (kPa)}

\item{gb}{Boundary layer conductance (mol/m2/s)}

\item{angles}{Named list with parameters for the leaf angle distribution (see documentation)}

\item{Ncmin}{Minimum leaf nitrogen content (g N/g DW)}

\item{sigma_soil}{Albedo of the soil surface}

\item{sky}{List with parameters for the sky model (see documentation)}

\item{lat}{Latitude of the location in degrees (positive values for Northern hemisphere)}

\item{Ca}{Air CO2 molar ratio (umol/mol)}
}
\value{
A Julia object with all the parameter values (access via $ notation)
}
\description{
Specify parameters of the virtual crop
}
\details{
The argument \code{angles} should be a named list with components \code{X} and\code{rtol}, the
first one being the parameter of the ellipsoidal leaf angle distribution model (a positive value)
and the second being the relative tolerance used by the adaptive cubature algorithm used to
calculated CO2 assimilation in the sunlit leaves.

The argument \code{sky} should be a named list with components \code{mode} and \code{nsky},
the first one being the string \code{"standard"} or \code{"uniform"} depending on the type of
sky model to be used and \code{nsky} being the number of hemispherical rings into which the
sky should be divided.
}
\examples{
pars = parameters()
pars$lat
}
