% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{toRad}
\alias{toRad}
\title{Convert hexadecimal degrees to radians}
\usage{
toRad(x)
}
\arguments{
\item{x}{Angle in hexadecimal degrees}
}
\value{
Angle in radians
}
\description{
Convert hexadecimal degrees to radians
}
\examples{
toRad(52)
}
