% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_interface.R
\name{variables}
\alias{variables}
\title{Specify the variables of the virtual crop}
\usage{
variables(
  pars,
  nleaf = 4,
  wleaf = 100,
  SLA = 300,
  kN = 0.4,
  f_Ncmn = 0.1,
  f_Ncmx = 0.3,
  pf_Nc = 4,
  f_Nrmn = 0.2,
  f_Nrmx = 0.6,
  pf_Nr = 2
)
}
\arguments{
\item{pars}{Parameters as generated by the \code{parameters}}

\item{nleaf}{Canopy nitrogen content (g N/m2)}

\item{wleaf}{Canopy (leaf) biomass as dry weight (g/m2)}

\item{SLA}{Specific leaf area (cm2/g)}

\item{kN}{Extinction coefficient of leaf nitrogen within the canopy}

\item{f_Ncmn}{Minimum fraction of N allocated to chlorophyll}

\item{f_Ncmx}{Maximum fraction of N allocated to chlorophyll}

\item{pf_Nc}{Scaling constant for fraction of N allocated to chlorophyll}

\item{f_Nrmn}{Minimum fraction of N allocated to Rubisco}

\item{f_Nrmx}{Maximum fraction of N allocated to Rubisco}

\item{pf_Nr}{Scaling constant for fraction of N allocated to Rubisco}
}
\value{
A Julia object with all the variables required to perform the simulations:
\describe{
\item{Nmin}{Minimum leaf nitrogen content (mol/m2)}
\item{GLAI}{Green leaf area index (limited by \code{nleaf} or \code{wleaf})}
\item{Nlmax}{Leaf nitrogen content on top of the canopy (mol/m2)}
\item{kN}{Extinction coefficient of leaf nitrogen within the canopy}
\item{f_Ncmn}{Minimum fraction of N allocated to chlorophyll}
\item{f_Ncmx}{Maximum fraction of N allocated to chlorophyll}
\item{pf_Nc}{Scaling constant for fraction of N allocated to chlorophyll}
\item{f_Nrmn}{Minimum fraction of N allocated to Rubisco}
\item{f_Nrmx}{Maximum fraction of N allocated to Rubisco}
\item{pf_Nr}{Scaling constant for fraction of N allocated to Rubisco}
}
}
\description{
Specify the variables of the virtual crop
}
\examples{
pars = parameters()
vars = variables(pars)
vars$GLAI
}
